fix(pathpoint): address review comments
