+++ integration_test_utilities/integration_test_mappings.yaml
@@ -510,10 +510,10 @@ resource_newrelic_workload_test.go:
   product_mapping: WORKLOADS
 resource_newrelic_pathpoint_flow.go:
   test: false
-  product_mapping: PATHPOINTS
+  product_mapping: PATHPOINT
 resource_newrelic_pathpoint_flow_test.go:
   test: true
-  product_mapping: PATHPOINTS
+  product_mapping: PATHPOINT
 structures_newrelic_alert_channel.go:
   test: false
   product_mapping: ALERTS_DEPRECATED
@@ -666,7 +666,7 @@ structures_newrelic_workload_test.go:
   product_mapping: WORKLOADS
 structures_newrelic_pathpoint.go:
   test: false
-  product_mapping: PATHPOINTS
+  product_mapping: PATHPOINT
 structures_newrelic_pathpoint_test.go:
   test: true
-  product_mapping: PATHPOINTS
+  product_mapping: PATHPOINT
