+++ newrelic/structures_newrelic_pathpoint.go
@@ -1,10 +1,538 @@
 package newrelic
 
 import (
+	"context"
+	"encoding/base64"
+	"fmt"
+	"strconv"
+	"strings"
+
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/newrelic/newrelic-client-go/v2/pkg/nrtime"
 	"github.com/newrelic/newrelic-client-go/v2/pkg/pathpoint"
 )
+
+// ── enum value slices (derived from Go client constants) ─────────────────────
+
+func pathpointFlowHealthRollupValues() []string {
+	return []string{
+		string(pathpoint.PathPointFlowHealthRollupTypes.ALERT_CONDITIONS),
+		string(pathpoint.PathPointFlowHealthRollupTypes.AUTOMATIC_ROLL_UP),
+	}
+}
+
+func pathpointRefreshIntervalValues() []string {
+	return []string{
+		string(pathpoint.PathPointRefreshIntervalTypes.ONE_MINUTE),
+		string(pathpoint.PathPointRefreshIntervalTypes.FIVE_MINUTES),
+		string(pathpoint.PathPointRefreshIntervalTypes.TEN_MINUTES),
+		string(pathpoint.PathPointRefreshIntervalTypes.FIFTEEN_MINUTES),
+		string(pathpoint.PathPointRefreshIntervalTypes.THIRTY_MINUTES),
+	}
+}
+
+func pathpointKpiAggregationValues() []string {
+	return []string{
+		string(pathpoint.PathPointKpiNRQLAggregationsTypes.AVERAGE),
+		string(pathpoint.PathPointKpiNRQLAggregationsTypes.COUNT),
+		string(pathpoint.PathPointKpiNRQLAggregationsTypes.HISTOGRAM),
+		string(pathpoint.PathPointKpiNRQLAggregationsTypes.MAX),
+		string(pathpoint.PathPointKpiNRQLAggregationsTypes.MIN),
+		string(pathpoint.PathPointKpiNRQLAggregationsTypes.PERCENTILE),
+		string(pathpoint.PathPointKpiNRQLAggregationsTypes.SUM),
+		string(pathpoint.PathPointKpiNRQLAggregationsTypes.UNIQUE_COUNT),
+	}
+}
+
+func pathpointKpiTimeDurationValues() []string {
+	return []string{
+		string(pathpoint.PathPointKpiTimeDurationTypes.SEVEN_DAYS),
+		string(pathpoint.PathPointKpiTimeDurationTypes.SIXTY_MINUTES),
+		string(pathpoint.PathPointKpiTimeDurationTypes.SIX_HOURS),
+		string(pathpoint.PathPointKpiTimeDurationTypes.THIRTY_DAYS),
+		string(pathpoint.PathPointKpiTimeDurationTypes.THIRTY_MINUTES),
+		string(pathpoint.PathPointKpiTimeDurationTypes.THREE_HOURS),
+		string(pathpoint.PathPointKpiTimeDurationTypes.TWENTY_FOUR_HOURS),
+	}
+}
+
+func pathpointSignalTypeValues() []string {
+	return []string{
+		string(pathpoint.PathPointSignalTypeTypes.ENTITY),
+		string(pathpoint.PathPointSignalTypeTypes.ALERT),
+	}
+}
+
+func pathpointStepHealthRollupValues() []string {
+	return []string{
+		string(pathpoint.PathPointStepHealthRollupTypes.BEST_STATUS_WINS),
+		string(pathpoint.PathPointStepHealthRollupTypes.WORST_STATUS_WINS),
+	}
+}
+
+func pathpointThresholdTypeValues() []string {
+	return []string{
+		string(pathpoint.PathPointThresholdTypeTypes.FIXED),
+		string(pathpoint.PathPointThresholdTypeTypes.PERCENTAGE),
+	}
+}
+
+func pathpointStageHealthRollupValues() []string {
+	return []string{
+		string(pathpoint.PathPointStageHealthRollupTypes.ALERT_CONDITIONS),
+		string(pathpoint.PathPointStageHealthRollupTypes.AUTOMATIC_ROLL_UP),
+	}
+}
+
+// ── schema definitions ────────────────────────────────────────────────────────
+
+func pathpointFlowSchema() map[string]*schema.Schema {
+	kpiQuerySelectSchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"aggregation_type": {
+				Type:         schema.TypeString,
+				Required:     true,
+				Description:  "Aggregation function: AVERAGE, COUNT, HISTOGRAM, MAX, MIN, PERCENTILE, SUM, UNIQUE_COUNT.",
+				ValidateFunc: validation.StringInSlice(pathpointKpiAggregationValues(), false),
+			},
+			"alias": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Optional alias for the aggregated value.",
+			},
+			"attribute": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Attribute name to aggregate. Required for all functions except COUNT.",
+			},
+			"threshold": {
+				Type:        schema.TypeFloat,
+				Optional:    true,
+				Description: "Threshold used in the selected function.",
+			},
+		},
+	}
+
+	kpiRelativeRangeSchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"since": {
+				Type:         schema.TypeString,
+				Required:     true,
+				Description:  "How far back the KPI is evaluated.",
+				ValidateFunc: validation.StringInSlice(pathpointKpiTimeDurationValues(), false),
+			},
+			"compare_against": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Description:  "The earlier window to compare against.",
+				ValidateFunc: validation.StringInSlice(pathpointKpiTimeDurationValues(), false),
+			},
+		},
+	}
+
+	kpiQuerySchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"from": {
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "Data source to query from (e.g., Transaction, Metric, Log).",
+			},
+			"where": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Optional WHERE clause to filter data.",
+			},
+			"select": {
+				Type:        schema.TypeList,
+				Required:    true,
+				MaxItems:    1,
+				Description: "SELECT clause defining what to aggregate.",
+				Elem:        kpiQuerySelectSchema,
+			},
+			"time_window": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MaxItems:    1,
+				Description: "Time window for KPI evaluation.",
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"custom_range": {
+							Type:        schema.TypeString,
+							Optional:    true,
+							Description: "Raw NRQL time fragment, e.g. 'SINCE 3 days ago COMPARE WITH 1 day ago'. Mutually exclusive with relative_range.",
+						},
+						"relative_range": {
+							Type:        schema.TypeList,
+							Optional:    true,
+							MaxItems:    1,
+							Description: "Relative time window. Mutually exclusive with custom_range.",
+							Elem:        kpiRelativeRangeSchema,
+						},
+					},
+				},
+			},
+		},
+	}
+
+	kpiSchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"id": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "The unique identifier of the KPI.",
+			},
+			"name": {
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "Display name of the KPI.",
+			},
+			"description": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Optional description.",
+			},
+			"category": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Optional category to group KPIs.",
+			},
+			"account_id": {
+				Type:        schema.TypeInt,
+				Optional:    true,
+				Computed:    true,
+				Description: "Account ID this KPI belongs to. Defaults to the flow's account_id.",
+			},
+			"query": {
+				Type:        schema.TypeList,
+				Required:    true,
+				MaxItems:    1,
+				Description: "NRQL query definition for this KPI.",
+				Elem:        kpiQuerySchema,
+			},
+			"metric_query": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "NRQL query using Metric, derived after processing event-to-metric rules. Read-only.",
+			},
+		},
+	}
+
+	signalSchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"guid": {
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "Entity GUID of the signal.",
+			},
+			"name": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Computed:    true,
+				Description: "Display name of the signal.",
+			},
+			"type": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Description:  "Whether this GUID belongs to an entity or an alert condition: ENTITY or ALERT.",
+				ValidateFunc: validation.StringInSlice(pathpointSignalTypeValues(), false),
+			},
+			"is_excluded": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+				Description: "When true, this signal is excluded from step health calculation.",
+			},
+		},
+	}
+
+	stepConfigSchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"health_rollup": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Description:  "How step health is rolled up: BEST_STATUS_WINS or WORST_STATUS_WINS.",
+				ValidateFunc: validation.StringInSlice(pathpointStepHealthRollupValues(), false),
+			},
+			"threshold_type": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Description:  "Whether threshold is FIXED or PERCENTAGE.",
+				ValidateFunc: validation.StringInSlice(pathpointThresholdTypeValues(), false),
+			},
+			"threshold_value": {
+				Type:        schema.TypeInt,
+				Optional:    true,
+				Description: "Numeric threshold value for step health evaluation.",
+			},
+		},
+	}
+
+	stepSchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"id": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "Internal step workload ID, used for updates.",
+			},
+			"name": {
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "Display name of the step.",
+			},
+			"is_excluded": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+				Description: "When true, this step is excluded from level health calculation.",
+			},
+			"link": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Optional URL to an external resource.",
+			},
+			"scoped_accounts": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MinItems:    1,
+				Description: "Account IDs whose data is scoped to this step.",
+				Elem:        &schema.Schema{Type: schema.TypeInt},
+			},
+			"entity_search_query": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MaxItems:    1,
+				Description: "Filter query used to fetch signals for this step.",
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"query": {
+							Type:        schema.TypeString,
+							Required:    true,
+							Description: "Filter query for signals, e.g. domain='NR1' AND type='APPLICATION'.",
+						},
+						"is_excluded": {
+							Type:        schema.TypeBool,
+							Optional:    true,
+							Default:     false,
+							Description: "When true, this query is excluded from health calculation.",
+						},
+					},
+				},
+			},
+			"config": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MaxItems:    1,
+				Description: "Health evaluation configuration for this step.",
+				Elem:        stepConfigSchema,
+			},
+			"signals": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MinItems:    1,
+				Description: "Entity signals associated with this step.",
+				Elem:        signalSchema,
+			},
+		},
+	}
+
+	levelSchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"id": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "Internal level workload ID, used for updates.",
+			},
+			"steps": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MinItems:    1,
+				MaxItems:    50,
+				Description: "Ordered list of steps within this level.",
+				Elem:        stepSchema,
+			},
+		},
+	}
+
+	relatedSchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"source": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+				Description: "When true, this stage acts as a source to other stages.",
+			},
+			"target": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+				Description: "When true, this stage acts as a target to other stages.",
+			},
+		},
+	}
+
+	stageSchema := &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"id": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "Internal stage workload ID, used for updates.",
+			},
+			"name": {
+				Type:        schema.TypeString,
+				Required:    true,
+				Description: "Display name of the stage.",
+			},
+			"health_rollup": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				Description:  "Health rollup strategy: ALERT_CONDITIONS or AUTOMATIC_ROLL_UP.",
+				ValidateFunc: validation.StringInSlice(pathpointStageHealthRollupValues(), false),
+			},
+			"is_excluded": {
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+				Description: "When true, this stage is excluded from flow health calculation.",
+			},
+			"link": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Optional URL to an external resource.",
+			},
+			"related": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MaxItems:    1,
+				Description: "Relationship role of this stage within the flow.",
+				Elem:        relatedSchema,
+			},
+			"stage_kpis": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MinItems:    1,
+				Description: "KPIs tracked at the stage level.",
+				Elem:        kpiSchema,
+			},
+			"levels": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				MinItems:    1,
+				MaxItems:    50,
+				Description: "Ordered list of levels within this stage.",
+				Elem:        levelSchema,
+			},
+		},
+	}
+
+	return map[string]*schema.Schema{
+		"account_id": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "The New Relic account ID that owns this Pathpoint flow.",
+		},
+		"name": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "Display name of the Pathpoint flow.",
+		},
+		"description": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Optional description of the flow.",
+		},
+		"category": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Optional category used to group flows (e.g. Marketing, Checkout).",
+		},
+		"health_rollup": {
+			Type:         schema.TypeString,
+			Optional:     true,
+			Description:  "Health rollup strategy: ALERT_CONDITIONS or AUTOMATIC_ROLL_UP.",
+			ValidateFunc: validation.StringInSlice(pathpointFlowHealthRollupValues(), false),
+		},
+		"refresh_interval": {
+			Type:         schema.TypeString,
+			Optional:     true,
+			Description:  "How often health statuses refresh: ONE_MINUTE, FIVE_MINUTES, TEN_MINUTES, FIFTEEN_MINUTES, THIRTY_MINUTES.",
+			ValidateFunc: validation.StringInSlice(pathpointRefreshIntervalValues(), false),
+		},
+		"kpis": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			MinItems:    1,
+			Description: "KPIs tracked at the flow level.",
+			Elem:        kpiSchema,
+		},
+		"stages": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			MinItems:    1,
+			MaxItems:    50,
+			Description: "Ordered list of stages that make up this flow.",
+			Elem:        stageSchema,
+		},
+		"guid": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "The entity GUID assigned to this Pathpoint flow.",
+		},
+		"version": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Last updated timestamp, used for version control.",
+		},
+	}
+}
+
+// ── import helper ─────────────────────────────────────────────────────────────
+
+// resourceNewRelicPathpointFlowImport supports two import ID formats:
+//   - "<guid>"            — account_id is decoded from the GUID
+//   - "<account_id>:<guid>" — explicit account_id provided by the user
+func resourceNewRelicPathpointFlowImport(ctx context.Context, d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
+	parts := strings.SplitN(d.Id(), ":", 2)
+	if len(parts) == 2 {
+		accountID, err := strconv.Atoi(parts[0])
+		if err != nil {
+			return nil, fmt.Errorf("invalid account_id in import ID %q: %w", d.Id(), err)
+		}
+		_ = d.Set("account_id", accountID)
+		d.SetId(parts[1])
+	} else {
+		accountID := accountIDFromGUID(d.Id())
+		if accountID == 0 {
+			return nil, fmt.Errorf("cannot determine account_id from GUID %q; use 'account_id:guid' import format", d.Id())
+		}
+		_ = d.Set("account_id", accountID)
+	}
+
+	diags := resourceNewRelicPathpointFlowRead(ctx, d, meta)
+	if diags.HasError() {
+		return nil, fmt.Errorf("error reading Pathpoint flow during import: %s", diags[0].Summary)
+	}
+	return []*schema.ResourceData{d}, nil
+}
+
+// accountIDFromGUID decodes a New Relic entity GUID (base64 of "accountID|domain|type|id")
+// and returns the account ID. Returns 0 if the GUID cannot be parsed.
+func accountIDFromGUID(guid string) int {
+	decoded, err := base64.StdEncoding.DecodeString(guid)
+	if err != nil {
+		return 0
+	}
+	parts := strings.SplitN(string(decoded), "|", 2)
+	if len(parts) < 1 {
+		return 0
+	}
+	id, err := strconv.Atoi(parts[0])
+	if err != nil {
+		return 0
+	}
+	return id
+}
 
 // ── expand helpers (Terraform state → API input) ──────────────────────────────
